Replace astroconda in Setup Notebook
The Setup Notebook is the crucial Notebook to alter to remove the dependency on astroconda, which is currently not working with Python>3.7 .
We need Python >=3.8 for CalCOS to run, so we want to alter the Notebooks to use a non astroconda environment, at least for the time being.
Since only 10 packages are required for all the existing Notebooks, we opted to just explain how the user should explicitly download them with `conda` and `pip`.