Generalize algorithm for crowded scatterplots
The algorithm used to make the scatterplot for this notebook includes a
"magic" value constant plus a random number generator to prune the
plotted points in regions where they overlap.  That makes the plot
faster and reduces the size of the plot files (for some plot types,
e.g., pdf).

This version includes a new function select_subset() that calculates the
sampling parameters (no need for a magic hand-tuned value).  It also
eliminates the random sampling in favor of a simpler regular sampling of
the distribution. The results are visually indistinguishable from the
previous version.

Note that the select_subset() function gets called after the
plt.figure() call so that it can use the plot dimensions to determine
the symbol overlap properties.

This version also replaces the RectBivariateSpline interpolator with
RegularGridInterpolator, which eliminates ringing in regions where the
distribution changes rapidly.  That does not affect the resulting plot
appearance but makes the calculation more robust.